refactor(12.0): Update Difficulty enum type from uint8 to int16
TrinityCore 12.0 changed the core Difficulty enum from uint8 to int16.
This commit updates all Playerbot module enums and function signatures
to match:

- InstanceFarmingManager.h: InstanceDifficulty enum
- RaidState.h: RaidDifficulty enum
- DungeonState.h: DungeonDifficulty enum
- GroupEvents.h/cpp: DifficultyChanged() parameter
- PlayerbotGroupScript.h: GroupStateSnapshot difficulty fields
- PlayerbotGroupScript.cpp: Difficulty casts
- PlayerBotHooks.cpp: GroupEvent difficulty cast

Also adds comprehensive WOW_12_0_MIGRATION_ANALYSIS.md documenting:
- Critical API changes (Stats, Difficulty, Items, Spells)
- Files requiring updates
- Migration task plan

+++ src/modules/Playerbot/AI/Coordination/Dungeon/DungeonState.h
@@ -38,8 +38,9 @@ enum class DungeonState : uint8
 /**
  * @enum DungeonDifficulty
  * @brief Dungeon difficulty modes
+ * @note WoW 12.0: Changed from uint8 to int16 to match core Difficulty type
  */
-enum class DungeonDifficulty : uint8
+enum class DungeonDifficulty : int16
 {
     NORMAL = 0,
     HEROIC = 1,
